+++ frontend/package.json
@@ -20,6 +20,7 @@
     "jspdf": "^3.0.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-markdown": "^10.1.0",
     "react-router": "^7.1.5",
     "react-router-dom": "^7.4.1",
     "tailwindcss": "^4.0.4"
