+++ frontend/src/main.jsx
@@ -1,6 +1,3 @@
-
-
-
 /**
  * @module main.jsx
  * This is the entry point for the React application. It initializes the React DOM rendering,
@@ -8,14 +5,10 @@
  * the Chat page, and a fallback 404 page.
  */
 
-
-
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router";
 import App from "./App.jsx";
-import Chat from "./Pages/chat.jsx";
 import PageNotFound from "./Pages/404.jsx";
-
 
 const root = document.getElementById("root");
 /**
@@ -30,7 +23,6 @@ ReactDOM.createRoot(root).render(
 	<BrowserRouter>
 		<Routes>
 			<Route path='/' element={<App />} />
-			<Route path='/chat' element={<Chat />} />
 			<Route path='*' element={<PageNotFound />} />
 		</Routes>
 	</BrowserRouter>
